requirements.txt
absl-py==1.0.0
anyio==3.5.0
asgiref==3.5.1
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
black==22.1.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
debugpy==1.5.1
decorator==5.1.1
efficientnet==1.1.1
executing==0.8.2
fastapi==0.73.0
flatbuffers==2.0
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.0
h11==0.13.0
h5py==2.10.0
idna==3.3
imageio==2.19.0
importlib-metadata==4.11.3
ipykernel==6.9.1
ipython==8.0.1
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.2
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.3.6
matplotlib-inline==0.1.3
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mypy-extensions==0.4.3
nest-asyncio==1.5.4
networkx==2.8
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.1
prompt-toolkit==3.0.28
protobuf==3.20.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
Pygments==2.11.2
pyparsing==3.0.8
python-multipart==0.0.5
PyWavelets==1.3.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scipy==1.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.2.0
stack-data==0.2.0
starlette==0.17.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tifffile==2022.5.4
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.4
uvloop==0.16.0
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
